Implemented draft of dataset removal in S3, except missing `get_all_datasets` piece of the puzzle
