+++ kamu-core/tests/tests/repos/test_dataset_repository_shared.rs
@@ -130,3 +130,50 @@ pub async fn test_rename_dataset(repo: &dyn DatasetRepository) {
     use futures::StreamExt;
     assert_eq!(baz.as_metadata_chain().iter_blocks().count().await, 2);
 }
+
+/////////////////////////////////////////////////////////////////////////////////////////
+
+pub async fn test_delete_dataset(repo: &dyn DatasetRepository) {
+    let name_foo = DatasetName::new_unchecked("foo");
+    let name_bar = DatasetName::new_unchecked("bar");
+
+    let snapshots = vec![
+        MetadataFactory::dataset_snapshot()
+            .name("foo")
+            .kind(DatasetKind::Root)
+            .push_event(MetadataFactory::set_polling_source().build())
+            .build(),
+        MetadataFactory::dataset_snapshot()
+            .name("bar")
+            .kind(DatasetKind::Derivative)
+            .push_event(MetadataFactory::set_transform(["foo"]).build())
+            .build(),
+    ];
+
+    let handles: Vec<_> = create_datasets_from_snapshots(repo, snapshots)
+        .await
+        .into_iter()
+        .map(|(_, r)| r.unwrap().dataset_handle)
+        .collect();
+
+    assert_matches!(
+        repo.delete_dataset(&name_foo.as_local_ref()).await,
+        Err(DeleteDatasetError::DanglingReference(e)) if e.children == vec![handles[1].clone()]
+    );
+
+    assert!(repo.get_dataset(&name_foo.as_local_ref()).await.is_ok());
+    assert!(repo.get_dataset(&name_bar.as_local_ref()).await.is_ok());
+
+    repo.delete_dataset(&name_bar.as_local_ref()).await.unwrap();
+    repo.delete_dataset(&name_foo.as_local_ref()).await.unwrap();
+
+    assert_matches!(
+        repo.get_dataset(&name_foo.as_local_ref())
+            .await
+            .err()
+            .unwrap(),
+        GetDatasetError::NotFound(_),
+    )
+}
+
+/////////////////////////////////////////////////////////////////////////////////////////
